Detect tech/warehouse_vision_system/requirements.txt
# Core Dependencies
numpy>=1.21.0
opencv-python>=4.5.0
torch>=1.12.0
torchvision>=0.13.0
onnx>=1.12.0
tensorrt>=8.4.0

# NVIDIA Stack
pydeepstream>=6.0
tlt-toolkit>=5.0.0
cuda-python>=11.7.0

# Computer Vision
scikit-image>=0.19.0
scikit-learn>=1.1.0
scipy>=1.8.0
pillow>=9.0.0

# Data Processing
pandas>=1.4.0
h5py>=3.7.0
pyyaml>=6.0
jsonschema>=4.5.0

# Networking & Communication
redis>=4.3.0
pika>=1.3.0
requests>=2.28.0
websockets>=10.3

# Monitoring & Logging
prometheus-client>=0.14.0
grafana-api>=1.0.3
structlog>=22.1.0

# 3D Reconstruction
open3d>=0.15.0
pyvista>=0.36.0
trimesh>=3.12.0

# Deployment
docker>=6.0.0
kubernetes>=24.2.0

# Testing
pytest>=7.1.0
pytest-cov>=3.0.0
pytest-asyncio>=0.19.0
